requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
anthropic==0.49.0
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0
pypdf==5.4.0
python-multipart==0.0.20
supabase==2.15.0
tenacity==9.1.2